requirements.txt
Django==5.2.0
djangorestframework==3.16.0
django-cors-headers==4.7.0
Pillow==11.2.1
python-dotenv==1.0.1
pymysql==1.1.1
cryptography==44.0.2
celery==5.5.1
coreapi==2.3.3
pyyaml==6.0.2
redis==5.0.3